Move unit tests to unit folder
