--- tests/mocks/node-modules/mock-all.ts
@@ -1,4 +1,0 @@
-import 'tests/mocks/node-modules/koa2-swagger-ui';
-import 'tests/mocks/node-modules/fs';
-import 'tests/mocks/node-modules/js-yaml';
-import 'tests/mocks/node-modules/json-refs';
