+++ tests/unit/cases/src/controllers/health.test.ts
@@ -1,8 +1,8 @@
-import 'tests/mocks/services/health';
+import 'tests/unit/mocks/services/health';
 import { Context } from 'koa';
 import { createMockContext } from '@shopify/jest-koa-mocks';
 import HealthController from 'src/controllers/health';
-import healthInfoSchema from 'tests/schemas/health';
+import healthInfoSchema from 'tests/unit/schemas/health';
 
 describe('health controllers', (): void => {
   afterAll((): void => {
