+++ tests/unit/cases/src/controllers/login.test.ts
@@ -1,9 +1,9 @@
-import 'tests/mocks/services/login';
+import 'tests/unit/mocks/services/login';
 import { Context } from 'koa';
 import { createMockContext } from '@shopify/jest-koa-mocks';
 import LoginController from 'src/controllers/login';
-import loginSchema from 'tests/schemas/login';
-import errorSchema from 'tests/schemas/error';
+import loginSchema from 'tests/unit/schemas/login';
+import errorSchema from 'tests/unit/schemas/error';
 
 describe('login controller', (): void => {
   afterAll((): void => {
