+++ tests/unit/cases/src/middlewares/login.test.ts
@@ -1,8 +1,8 @@
-import 'tests/mocks/validators/login';
+import 'tests/unit/mocks/validators/login';
 import { createMockContext } from '@shopify/jest-koa-mocks';
 import { Context } from 'koa';
 import LoginMiddleware from 'src/middlewares/login';
-import errorSchema from 'tests/schemas/error';
+import errorSchema from 'tests/unit/schemas/error';
 
 describe('login middleware', (): void => {
   afterAll((): void => {
