+++ tests/unit/cases/src/models/responses/error.test.ts
