+++ tests/unit/cases/src/settings/load.test.ts
