+++ tests/unit/mocks/node-modules/fs.ts
