+++ tests/unit/mocks/node-modules/js-yaml.ts
