+++ tests/unit/mocks/node-modules/json-refs.ts
