+++ tests/unit/mocks/node-modules/mock-all.ts
@@ -1,0 +1,4 @@
+import 'tests/unit/mocks/node-modules/koa2-swagger-ui';
+import 'tests/unit/mocks/node-modules/fs';
+import 'tests/unit/mocks/node-modules/js-yaml';
+import 'tests/unit/mocks/node-modules/json-refs';
