+++ tests/unit/schemas/login.ts
@@ -1,4 +1,4 @@
-import 'tests/matchers/to-be-types';
+import 'tests/unit/matchers/to-be-types';
 
 export default {
   username: expect.any(String),
